+++ CMakeLists.txt
@@ -6,37 +6,37 @@ set(CMAKE_CXX_FLAGS "-pthread -std=c++17 -Wall")
 
 add_executable(nsfd-server
         src/nsfd.cc
-        src/server/SocketFactory.cc
-        src/server/SocketFactory.h
-        src/server/HealthCheckServer.cc
-        src/server/HealthCheckServer.h
-		src/server/TcpServer.cc
-        src/server/TcpServer.h
-		src/server/Socket.cc
-		src/server/Socket.h
-		src/server/api/HealthCheckRequest.h
-		src/server/api/HealthCheckRequest.cc
-		src/server/ServerSupervisor.h
-		src/server/ServerSupervisor.cc
-		src/server/api/HealthCheckRequestDeserializer.h
-		src/server/api/HealthCheckRequestDeserializer.cc
-		src/network/IcmpClient.h
-		src/network/IcmpClient.cc
+		src/server/socket_factory.cc
+		src/server/socket_factory.h
+		src/server/health_check_server.cc
+		src/server/health_check_server.h
+		src/server/tcp_server.cc
+		src/server/tcp_server.h
+		src/server/socket.cc
+		src/server/socket.h
+		src/server/api/health_check_request.h
+		src/server/api/health_check_request.cc
+		src/server/server_supervisor.h
+		src/server/server_supervisor.cc
+		src/server/api/health_check_request_deserializer.h
+		src/server/api/health_check_request_deserializer.cc
+		src/network/icmp_client.h
+		src/network/icmp_client.cc
 		src/network/dns.h
 		src/network/dns.cc
-		src/network/TcpClient.h
-		src/network/TcpClient.cc
-		src/health/HealthCheckTask.h
-		src/health/HealthCheckTask.cc
-		src/health/HealthCheckTaskScheduler.h
-		src/health/HealthCheckTaskScheduler.cc
-		src/health/HealthCheckTaskStorage.h
-		src/health/HealthCheckTaskStorage.cc
-		src/health/HealthCheckTaskFactory.h
-		src/health/HealthCheckTaskFactory.cc
-		src/util/CmdColor.h
-		src/server/MetricsServer.h
-		src/server/MetricsServer.cc
-		src/server/api/MetricsRequest.h
-		src/server/api/MetricsRequestDeserializer.h
-		src/server/api/MetricsRequestDeserializer.cc)
+		src/network/tcp_client.h
+		src/network/tcp_client.cc
+		src/health/health_check_task.h
+		src/health/health_check_task.cc
+		src/health/health_check_task_scheduler.h
+		src/health/health_check_task_scheduler.cc
+		src/health/health_check_task_storage.h
+		src/health/health_check_task_storage.cc
+		src/health/health_check_task_factory.h
+		src/health/health_check_task_factory.cc
+		src/util/cmd_colors.h
+		src/server/metrics_server.h
+		src/server/metrics_server.cc
+		src/server/api/metrics_request.h
+		src/server/api/metrics_request_deserializer.h
+		src/server/api/metrics_request_deserializer.cc)
