+++ src/health/health_check_task_storage.cc
@@ -1,4 +1,4 @@
-#include "HealthCheckTaskStorage.h"
+#include "health_check_task_storage.h"
 
 HealthCheckTaskStorage::~HealthCheckTaskStorage()
 {
