+++ src/health/health_check_task_storage.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <functional>
-#include "HealthCheckTask.h"
+#include "health_check_task.h"
 
 class HealthCheckTaskStorage
 {
