+++ src/server/api/metrics_request_deserializer.cc
@@ -1,4 +1,4 @@
-#include "MetricsRequestDeserializer.h"
+#include "metrics_request_deserializer.h"
 
 MetricsRequest* MetricsRequestDeserializer::deserialize(char *raw_message)
 {
