+++ src/server/metrics_server.h
@@ -1,8 +1,8 @@
 #ifndef NSFD_SERVER_METRICSSERVER_H_
 #define NSFD_SERVER_METRICSSERVER_H_
 
-#include "TcpServer.h"
-#include "../health/HealthCheckTaskStorage.h"
+#include "tcp_server.h"
+#include "../health/health_check_task_storage.h"
 
 class MetricsServer : public TcpServer
 {
