+++ src/server/socket.cc
@@ -1,4 +1,4 @@
-#include "Socket.h"
+#include "socket.h"
 
 #include <stdio.h>
 #include <string.h>
