+++ src/server/socket_factory.cc
@@ -1,6 +1,6 @@
 #include <arpa/inet.h>
 #include <cstring>
-#include "SocketFactory.h"
+#include "socket_factory.h"
 
 
 int SocketFactory::create_socket(int port) {
